+++ r5dev/mathlib/sha1.h
@@ -35,9 +35,9 @@ private:
     typedef unsigned long int uint32;   /* just needs to be at least 32bit */
     typedef unsigned long long uint64;  /* just needs to be at least 64bit */
  
-    static const unsigned int DIGEST_INTS = 5;  /* number of 32bit integers per SHA1 digest */
-    static const unsigned int BLOCK_INTS = 16;  /* number of 32bit integers per SHA1 block */
-    static const unsigned int BLOCK_BYTES = BLOCK_INTS * 4;
+    static const size_t DIGEST_INTS = 5;  /* number of 32bit integers per SHA1 digest */
+    static const size_t BLOCK_INTS = 16;  /* number of 32bit integers per SHA1 block */
+    static const size_t BLOCK_BYTES = BLOCK_INTS * 4;
  
     uint32 digest[DIGEST_INTS];
     std::string buffer;
@@ -47,7 +47,7 @@ private:
     void transform(uint32 block[BLOCK_BYTES]);
  
     static void buffer_to_block(const std::string &buffer, uint32 block[BLOCK_BYTES]);
-    static void read(std::istream &is, std::string &s, int max);
+    static void read(std::istream &is, std::string &s, size_t max);
 };
  
 std::string sha1(const std::string &string);
